Revert "Normalize Storage ETag condition headers"
This reverts commit bf814c876c76e51d9ac0751fce011e28b3408fcd.
