+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/implementation/util/ChunkedDownloadUtils.java
@@ -144,8 +144,9 @@ public class ChunkedDownloadUtils {
         // We don't want to modify the user's object, so we'll create a duplicate and set the retrieved etag.
         return new BlobRequestConditions().setIfModifiedSince(requestConditions.getIfModifiedSince())
             .setIfUnmodifiedSince(requestConditions.getIfUnmodifiedSince())
-            // Preserve the service-returned representation in the copied conditions. StorageEtagPolicy restores the
-            // RFC-compliant representation at the HTTP boundary without mutating this customer-visible value.
+            // Preserve the original service-returned ETag value here. Normalizing via
+            // StorageImplUtils.toETagHeaderValue() would change the request wire format for retry/download locking and
+            // should be treated as a separate compatibility decision rather than part of this bug fix.
             .setIfMatch(etag)
             .setIfNoneMatch(requestConditions.getIfNoneMatch())
             .setLeaseId(requestConditions.getLeaseId());
