+++ sdk/storage/azure-storage-common/src/main/java/com/azure/storage/common/implementation/StorageImplUtils.java
@@ -171,7 +171,7 @@ public class StorageImplUtils {
      * @return An RFC 9110 conformant entity-tag, or the original value when no conversion is needed.
      */
     public static String toETagHeaderValue(String eTag) {
-        if (eTag == null || Constants.HeaderConstants.ETAG_WILDCARD.equals(eTag)) {
+        if (eTag == null || eTag.isEmpty() || Constants.HeaderConstants.ETAG_WILDCARD.equals(eTag)) {
             return eTag;
         }
