+++ sdk/storage/azure-storage-common/src/test/java/com/azure/storage/common/implementation/StorageImplUtilsTests.java
@@ -68,7 +68,7 @@ public class StorageImplUtilsTests {
     }
 
     private static Stream<Arguments> etagValues() {
-        return Stream.of(Arguments.of(null, null), Arguments.of("", "\"\""),
+        return Stream.of(Arguments.of(null, null), Arguments.of("", ""),
             Arguments.of(Constants.HeaderConstants.ETAG_WILDCARD, Constants.HeaderConstants.ETAG_WILDCARD),
             Arguments.of("0x8DABC", "\"0x8DABC\""), Arguments.of("\"0x8DABC\"", "\"0x8DABC\""),
             Arguments.of("W/\"0x8DABC\"", "W/\"0x8DABC\""), Arguments.of("\"\"", "\"\""));
